[API] Removed API proxy
Removed the api.js proxy since the Já API has a legacy feature that
allows you to pass the API key as a URL parameter instead of a
header. This makes it so the browser does not send a pre-flight
OPTIONS request which is what caused the issue, since the Já API
does not respond to those.